requirements.txt
aiogram>=3.4.0
python-dotenv>=1.0.1
deep-translator>=1.11.4
langdetect>=1.0.9
edge-tts>=6.1.9
aiosqlite>=0.20.0
fastapi>=0.110.0
uvicorn>=0.29.0
httpx>=0.27.0
